Apply Random Forest models

Apply Random Forest models to predict SAV cover and presence/absence, with optional post-hoc processing.

Usage

sav_model(
  dat,
  type = c("cover", "pa"),
  depth = NULL,
  fetch = NULL,
  substrate = NULL,
  secchi = NULL,
  limitation = NULL,
  vmax_par = list(intercept = 1.4, slope = 1.33)
)

Arguments

dat

data.frame|sf
A data.frame or a sf object containing some or all of the following columns:

depth_m: Numeric, depth in meters.

fetch_km: Numeric, fetch in kilometers.

secchi: Numeric Secchi depth in meters (post_hoc)

substrate: Binary (0 = absent, 1 = present), indicating substrate limitations. (post_hoc)

limitation: Binary (0 = absent, 1 = present), indicating user-supplied limitations. Additional columns will be ignored.

type

character vector, either "cover" or "pa"
Model type(s).

depth, fetch

character
Column specification for the predictors, see Details.

substrate, secchi, limitation

character
Column specification for post_hoc variables, see Details.

vmax_par

named list
intercept and slope of the equation from Chambers and Kalff (1985) to compute the maximum depth of plant colonization (Vmax), see Details below.

Value

A data frame (or a sf object) containing the input columns along with model predictions.

The prediction column names match the values specified in type followed by the suffix _pred, and contain the raw model outputs (i.e., without post-hoc adjustment).

Post-hoc adjusted predictions (see Details) are included in additional columns with the same names as the type values, but with the suffix _post_hoc.

If a column secchi is present, then two additional columns are returned: vmax and limitation_secchi, see details for further explanation.

Details

There are two sets of models available. The first set consists of models predicting the presence or absence of SAV, while the second set focuses on SAV cover. Each set includes three random forest models: one using depth as a predictor, another using fetch, and a third combining both variables. For further details, see Croft-White (2022).

The selected model for generating predictions is determined by the type argument, which specifies the output as either cover or presence-absence, depending on the available predictors. The required input variables—depth, fetch, substrate, secchi, limitation—must correspond to column names in dat; otherwise, an error is thrown. If neither 'depth' nor 'fetch' is explicitly provided, the function will attempt to infer them from the column names. Matching is case-insensitive and will detect 'depth_m', 'depth', 'fetch_km' and 'fetch'.

If secchi is provided, two additional columns are returned:

vmax: Predicted maximum colonization depth calculated using the Chambers and Kalff equation.

limitation_secchi: Logical column indicating light limitation. TRUE if vmax >= depth_m, indicating the site is light-limited; FALSE otherwise.

The regression parameters for the Chambers and Kalff equation can be adjusted via the vmax_par argument. By default, parameters from Model A (Quebec and international lakes) in Croft-White et al. (2022) are used. To apply Model B (Quebec lakes only), use:

vmax_par = list(intercept = 1.32, slope = 1.14)

The post-hoc treatment adjusts raw predictions by setting them to 0 wherever limitations are present. Possible limitation columns include substrate, limitation, and limitation_secchi (see descriptions above). These columns are treated as binary indicators: any value greater than 0 is interpreted as a limiting condition. If a limitation is detected, the corresponding prediction is set to 0 in the post-hoc adjusted output.

References

Croft-White, M.V., Tang, R., Gardner Costa, J., Doka, S.E., and Midwood, J. D. 2022. Modelling submerged aquatic vegetation presence and percent cover to support the development of a freshwater fish habitat management tool. Can. Tech. Rep. Fish. Aquat. Sci. 3497: vi + 30 p.

Chambers, P.A., and Kalff, J. 1985. Depth Distribution and Biomass of Submerged aquatic macrophyte communities in relation to secchi depth. Can. J. Fish. Aquat. Sci. 42: 701–709

Examples

# \donttest{

# basic usage 
sav_model(data.frame(depth = c(5, 10)))
#> ℹ Looking for depth and fetch in column names.
#> ℹ Found Depth in column names.
#> ℹ Using cover and pa with depth
#>   depth_m pa_pred cover_pred pa_post_hoc cover_post_hoc
#> 1       5       1   85.43167           1       85.43167
#> 2      10       0   21.14400           0        0.00000
sav_model(data.frame(depth = c(5, 10), fetch = c(1, 2)), type = "pa")
#> ℹ Looking for depth and fetch in column names.
#> ℹ Found Fetch and Depth in column names.
#> ℹ Using pa with depth+fetch
#>   depth_m fetch_km pa_pred pa_post_hoc
#> 1       5        1       0           0
#> 2      10        2       0           0

# using post-hoc treatment 
sav_model(
 data.frame(
  depth = c(5, 10, 5), 
  fetch = c(1, 2, 10), 
  secchi = c(1, 10, 10), 
  substrate = c(TRUE, TRUE, FALSE)
 )
)
#> ℹ Looking for depth and fetch in column names.
#> ℹ Found Fetch and Depth in column names.
#> ℹ Using cover and pa with depth+fetch
#>   depth_m fetch_km secchi limitation_secchi     vmax substrate pa_pred
#> 1       5        1      1             FALSE  1.76890      TRUE       0
#> 2      10        2     10              TRUE 19.29351      TRUE       0
#> 3       5       10     10              TRUE 19.29351     FALSE       1
#>   cover_pred pa_post_hoc cover_post_hoc
#> 1   77.11233           0              0
#> 2   44.66533           0              0
#> 3   60.24833           0              0
# }